feat(ErrorViewer) : add the ErrorViewer class for error case
Add $ErrorViewer dealing with error case:
- 에러 케이스를 다루는 객체를 추가하였습니다.
